bib: rename "source" to "osinfo"
Now that we have the bib/internal/source package in the images
library the old name "source" is a bit too ambigious. Rename
to osinfo instead as this is what its about (maybe osinpect?
or osprobe?).
