+++ pkg/bib/container/solver.go
@@ -7,7 +7,7 @@ import (
 	"path/filepath"
 
 	"github.com/osbuild/images/pkg/arch"
-	"github.com/osbuild/images/pkg/bib/source"
+	"github.com/osbuild/images/pkg/bib/osinfo"
 	"github.com/osbuild/images/pkg/dnfjson"
 )
 
@@ -106,7 +106,7 @@ func (cnt *Container) setupRunSecrets() error {
 	return nil
 }
 
-func (cnt *Container) NewContainerSolver(cacheRoot string, architecture arch.Arch, sourceInfo *source.Info) (*dnfjson.Solver, error) {
+func (cnt *Container) NewContainerSolver(cacheRoot string, architecture arch.Arch, sourceInfo *osinfo.Info) (*dnfjson.Solver, error) {
 	solver := dnfjson.NewSolver(
 		sourceInfo.OSRelease.PlatformID,
 		sourceInfo.OSRelease.VersionID,
